Avoid reliance on logback classic
