+++ build.gradle
@@ -1,65 +1,3 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Java application project to get you started.
- * For more details take a look at the 'Building Java & JVM projects' chapter in the Gradle
- * User Manual available at https://docs.gradle.org/7.6/userguide/building_java_projects.html
- */
-
-plugins {
-    // Apply the application plugin to add support for building a CLI application in Java.
-    id 'application'
-    id 'eclipse'
-    id 'org.openjfx.javafxplugin' version '0.0.13'
-    id 'org.beryx.jlink' version '2.12.0'
-}
-
 tasks.named('wrapper') {
     distributionType = Wrapper.DistributionType.ALL
-}
-
-repositories {
-    // Use Maven Central for resolving dependencies.
-    mavenCentral()
-}
-
-dependencies {
-    // Use JUnit Jupiter for testing.
-    testImplementation 'org.junit.jupiter:junit-jupiter:5.9.1'
-
-    // Application dependencies
-    implementation 'org.slf4j:slf4j-api:2.0.4'
-
-    implementation 'ch.qos.logback:logback-classic:1.3.5'
-}
-
-println System.properties['java.home']
-
-java {
-    toolchain {
-        languageVersion = JavaLanguageVersion.of(17)
-    }
-}
-
-javafx {
-    version = "20"
-    modules = [ 'javafx.controls', 'javafx.fxml' ]
-}
-
-application {
-    mainModule = 'logviewer'
-    mainClass = 'io.github.qupath.logviewer.LogViewerApp'
-}
-
-tasks.named('test') {
-    // Use JUnit Platform for unit tests.
-    useJUnitPlatform()
-}
-
-
-jlink {
-    options = ['--strip-debug', '--compress', '2', '--no-header-files', '--no-man-pages']
-    launcher {
-        name = 'logviewer'
-    }
 }
