+++ log-viewer-api/src/main/java/io/github/qupath/logviewer/api/LoggerController.java
@@ -1,0 +1,5 @@
+package io.github.qupath.logviewer.api;
+
+public interface LoggerController {
+    void addLogMessage(LogMessage logMessage);
+}
