+++ log-viewer-api/src/main/java/module-info.java
@@ -1,0 +1,5 @@
+module logviewer.api {
+    exports io.github.qupath.logviewer.api;
+
+    requires org.slf4j;
+}
