+++ log-viewer-app/build.gradle
@@ -1,0 +1,55 @@
+plugins {
+    // Apply the application plugin to add support for building a CLI application in Java.
+    id 'application'
+
+    // Add JavaFX plugin.
+    id 'org.openjfx.javafxplugin' version '0.0.13'
+
+    // Add jlink plugin to add support for assembling modules into a custom runtime image.
+    id 'org.beryx.jlink' version '2.12.0'
+}
+
+repositories {
+    // Use Maven Central for resolving dependencies.
+    mavenCentral()
+}
+
+dependencies {
+    // Use JUnit Jupiter for testing.
+    testImplementation 'org.junit.jupiter:junit-jupiter:5.9.1'
+
+    // Use logging abstraction. This also import SLF4J.
+    implementation project(':log-viewer-api')
+
+    // Use implementation of Logback
+    implementation project(':log-viewer-logback')
+}
+
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(17)
+    }
+}
+
+javafx {
+    version = "20"
+    modules = [ 'javafx.controls', 'javafx.fxml' ]
+}
+
+application {
+    mainModule = 'logviewer'
+    mainClass = 'io.github.qupath.logviewer.app.LogViewerApp'
+}
+
+tasks.named('test') {
+    // Use JUnit Platform for unit tests.
+    useJUnitPlatform()
+}
+
+
+jlink {
+    options = ['--strip-debug', '--compress', '2', '--no-header-files', '--no-man-pages']
+    launcher {
+        name = 'logviewer'
+    }
+}
