+++ log-viewer-app/src/main/resources/io/github/qupath/logviewer/app/css/styles.css
