+++ log-viewer-app/src/main/resources/io/github/qupath/logviewer/app/strings_fr.properties
