+++ log-viewer-logback/build.gradle
@@ -1,0 +1,23 @@
+plugins {
+    // Apply the Java library plugin to add support for building a library in Java.
+    id 'java-library'
+}
+
+repositories {
+    // Use Maven Central for resolving dependencies.
+    mavenCentral()
+}
+
+dependencies {
+    // Use Logback as logging framework.
+    implementation 'ch.qos.logback:logback-classic:1.3.5'
+
+    // Use logging abstraction.
+    implementation project(':log-viewer-api')
+}
+
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(17)
+    }
+}
