+++ log-viewer-logback/src/main/java/module-info.java
@@ -1,0 +1,11 @@
+module logviewer.logback {
+    exports io.github.qupath.logviewer.logback;
+
+    requires org.slf4j;
+    requires ch.qos.logback.core;
+    requires ch.qos.logback.classic;
+    requires logviewer.api;
+
+    provides io.github.qupath.logviewer.api.LoggerManager with io.github.qupath.logviewer.logback.LogbackManager;
+
+}
