--- src/main/java/module-info.java
@@ -1,14 +1,0 @@
-module logviewer {
-
-    requires javafx.base;
-    requires javafx.controls;
-    requires javafx.fxml;
-
-    requires org.slf4j;
-    requires ch.qos.logback.core;
-    requires ch.qos.logback.classic;
-
-    opens io.github.qupath.logviewer to javafx.fxml;
-    exports io.github.qupath.logviewer to javafx.graphics;
-
-}
